requirements.txt
asgiref==3.3.1
Django==3.0
django-crispy-forms==1.11.2
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
gunicorn==20.0.4
Markdown==3.3.3
Pillow==8.1.0
psycopg2==2.8.6
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0
